Create the Todo class
As the user needs to be able to create numerous to-do's they need a way
to be created and stored. Defined a class to construct the todo objects.
Once the object is created, the only reason it needs to be accessed is
to display the data, allow the user to edit the data, and for deletion.
Set the class properties to be private and implemented getters and
setters to allow us to access the data and allow the user to edit the
data.

Additionally, imported the date-fns library for use with date validation
and manipulation.