New module to create storage arrays for todos
Workspaces will be the name of the storage areas for user todos. The
user will be able to create new workspaces and populate them with todo
objects.
The Workspace class also contains a private static array that stores all
new workspaces. This allows the user to modify the name of a workspace
or delete the workspace entirely.