Refactor workspaces for Single Responsiblity
Initial workspace class setup was responsible for managing both
workspace-instance functionality AND the static list of workspaces.
Refactored to create a WorkspaceManager. This manager is responsible for
the static list of workspaces, whilst the original Workspace class is
now only responsible for workspace instances.

Whilst refactoring, noticed that the deleteWorkspace method was set up
incorrectly. The workspace instances are objects containing a name and a
to-do list. deleteWorkspace was coded to look for a name directly,
rather than to look for a name inside an object. Modified code to
look for the name inside the objects in the static array.

However, by refactoring for Single Responsibility, the module now
has more conflict with other principles of SOLID code. Namely, the
dependency inversion principle, as WorkspaceManager depends on this
specific set up of Workspace.