New method to set current workspace + fix buttons
The idea for the app is that new todos are automatically added to the
currently selected workspace. To enable this functionality, we needed a
way to define the current workspace. Added this as a method to the
workspace manager.

With this new addition, was able to complete the button functions for
adding a new workspace and creating a new todo. When a new workspace is
created, it will automatically be set to the current workspace. When a
new todo is created, it's automatically added to the current workspace.