Add functionality for editing workspace details
User needs to be able to edit the details of a workspace. To do this,
created a separate form for editing a workspace. This form will be
prepopulated with the existing details of the selected workspace.

To access the edit form, added a nested span within the navbar button
for the workspace. Currently, the span is always visible, but in future
update will use CSS to make it so the span is only visible when hovering
over the button.

When the form is submitted, it will also refresh the nav bar, make the
edited workspace the currently selected workspace, and display the to-do
list of the workspace inside the workspace display.