Add checkbox to todo display to toggle complete
Each todo will now be displayed with an accompanying checkbox to toggle
whether the todo has been completed or not. Added functionality to this
checkbox to alter the class list of the to-do's parent div container so
that we can alter the styling of the to-do depending on whether it is
complete or not.