Add delete workspace functionality
Added a new button within the edit workspace dialog form. Clicking this
button will check whether the selected workspace is editable. If it is,
it will ask the user to confirm that they want to delete the workspace.
If the user clicks 'ok', it will delete the workspace, refresh the
workspace nav, and repopulate the workspace display.

Additional changes:
- Modified Workspace class so that it has a new private boolean value
'isEditable'. This is needed because we want there to always be a
default inbox available that the user is unable to edit or delete. By
having this new flag, we can check for this to dictate if the user can
edit or delete a workspace.
- created a new workspace-display-controller module in preparation for
adding additional functionality to the display. Moved
`populateWorkspaceDisplay` from the sidebar-controller script to this
new script as it's functionality is related to the display and not the
sidebar.
- Updated the create new workspace button to refresh the nav and
workspace display when a new workspace is created.
- Updated the create new to-do button to refresh the workspace display
when a new to-do is created.