Create additional function to refresh nav
When we refresh the nav, we also want to make sure the current workspace
is highlighted and that the nav items have the correct event listeners
implemented. These are all defined as individual functions, so created a
new function that calls the 3 individual functions to reduce code
elsewhere when refreshing the nav. Implemented this new function in
	- newWorkspaceSubmit
	- editWorkspaceSubmit
	- editWorkspaceDelete
	- createInboxWorkspace