Implement functionality to edit to-do's
Created a new dialog and form that will be opened when the user wants
to edit a to-do item. To access the new edit feature, populateWorkspace
has been modified to add an additional button to each to-do item.
Created additional functions to  prepopulate and submit this new form
inline with previous forms and submits.
This new button has been assigned a class of "todoChanges". This is so
we can utilise CSS in future to only display the buttons when the user
is hovering over a to-do item withinin the display.