Add functionality to move and delete to-do's
Within the to-do displays, added two additional buttons. One to delete
the specific to-do. The other to move it from its current workspace to
another pre-existing workspace.
Additionally, had to modify the function that adds event listeners to
the nav bar items. This is because in its initial set up it wasn't
- Updating which workspace is currently active in the display
- Correctly updating the display to show the current workspace to-do's.
To fix this, I've had to make the function more dependent on the
WorkspaceManager class.