Implement page styling
Implemented styling for
- dialogs
- forms
- form buttons
- todo display
- navbar buttons
- radio buttons

Additional changes:
- Implemented SVGs for the workspace edit, todo edit, move todo, and
delete todo buttons. Using in-line svg as opposed to svg files so that
the color of the svg can be controlled via css.
- Bug fix for the move todo select input. Initial set up didn't clear
the innerHTML of the select input meaning when the input got refreshed
it was storing duplicate entries for pre-existing workspaces. Fixed this
by clearning the options prior to refreshing them.