Add color option to workspaces
Improved the workspace class and forms to allow the user to select a
color for the workspace. Each workspace can have its own color and when
the workspace is active the button in the navbar as well as the display
background changes color to match the chosen color for the workspaces.