Fix bug duplicating workspace list on move todo
When moving a todo, the select input for the move location wasn't being
cleared. This meant everytime the user moved a todo it was appending
options for each workspace available. The more to-do's moved, the more
duplicate options were being added. Fixed by clearing the select input
innerHTML prior to repopulating with the current workspace list.